requirements.txt
flask
rembg
pillow
numpy
gunicorn
